Refactored lesson-1 and lesson-2
