--- lesson-2/channel_example_1.txt
@@ -1,1 +1,0 @@
-fegergefgru
