--- lesson-2/channel_example_3.txt
@@ -1,1 +1,0 @@
-New String to write to file...1610994460218
