--- lesson-2/channel_example_4.txt
@@ -1,9 +1,0 @@
-1234567890
-1234567890
-1234567890
-1234567890
-1234567890
-1234567890
-1234567890
-1234567890
-12
