+++ lesson-2/src/test/java/ru/daniilazarnov/MainTest.java
@@ -17,117 +17,8 @@ import static org.junit.jupiter.api.Assertions.*;
 class MainTest {
 
     @Test
-    void testWalk() throws IOException {
-        Stream<Path> stream = Files.walk(Path.of("../lesson-1/dir"));
-
-        stream.forEach(System.out::println);
-    }
-
-    @Test
-    void testFileVisitResult() {
-        for (FileVisitResult value : FileVisitResult.values()) {
-            System.out.println(value);
-        }
-    }
-
-    @Test
-    void testWalkFileTree() throws IOException {
-        Files.walkFileTree(Path.of("../lesson-1/dir"), new FileVisitor<Path>() {
-            @Override
-            public FileVisitResult preVisitDirectory(Path dir, BasicFileAttributes attrs) throws IOException {
-                return FileVisitResult.CONTINUE;
-            }
-
-            @Override
-            public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
-                return FileVisitResult.CONTINUE;
-            }
-
-            @Override
-            public FileVisitResult visitFileFailed(Path file, IOException exc) throws IOException {
-                return FileVisitResult.CONTINUE;
-            }
-
-            @Override
-            public FileVisitResult postVisitDirectory(Path dir, IOException exc) throws IOException {
-                return FileVisitResult.CONTINUE;
-            }
-        });
-    }
-
-    @Test
-    void testDeleteEmptyDirectory() throws IOException {
-        Files.createDirectory(Path.of("empty_dir"));
-
-        assertTrue(Files.exists(Path.of("empty_dir")));
-
-        Files.delete(Path.of("empty_dir"));
-
-        assertFalse(Files.exists(Path.of("empty_dir")));
-    }
-
-    @Test
-    void testDeleteNotEmptyDirectory() throws IOException {
-        Files.createDirectory(Path.of("dir_with_file"));
-        Files.createFile(Path.of("dir_with_file/file.txt"));
-
-        assertTrue(Files.exists(Path.of("dir_with_file")));
-        assertThrows(DirectoryNotEmptyException.class, () -> Files.delete(Path.of("dir_with_file")));
-
-        assertTrue(Files.exists(Path.of("dir_with_file")));
-        assertTrue(Files.exists(Path.of("dir_with_file/file.txt")));
-
-        Files.walkFileTree(Path.of("dir_with_file"), new FileVisitor<Path>() {
-            @Override
-            public FileVisitResult preVisitDirectory(Path dir, BasicFileAttributes attrs) throws IOException {
-                return FileVisitResult.CONTINUE;
-            }
-
-            @Override
-            public FileVisitResult visitFile(Path file, BasicFileAttributes attrs) throws IOException {
-                Files.delete(file);
-                return FileVisitResult.CONTINUE;
-            }
-
-            @Override
-            public FileVisitResult visitFileFailed(Path file, IOException exc) throws IOException {
-                return FileVisitResult.CONTINUE;
-            }
-
-            @Override
-            public FileVisitResult postVisitDirectory(Path dir, IOException exc) throws IOException {
-                Files.delete(dir);
-                return FileVisitResult.CONTINUE;
-            }
-        });
-
-        assertFalse(Files.exists(Path.of("dir_with_file")));
-        assertFalse(Files.exists(Path.of("dir_with_file/file.txt")));
-    }
-
-    @Test
-    void testFileChannel() throws IOException {
-        Path srcPath = Path.of("channel_example_1.txt");
-
-        RandomAccessFile src = new RandomAccessFile(srcPath.toFile(), "rw");
-        FileChannel srcChannel = src.getChannel();
-
-        System.out.println(Files.readString(srcPath));
-
-
-        RandomAccessFile dest = new RandomAccessFile(Path.of("channel_example_2.txt").toFile(), "rw");
-        FileChannel destChannel = dest.getChannel();
-
-        srcChannel.transferTo(0, srcChannel.size(), destChannel);
-
-        //DatagramChannel - UDP
-        //SocketChannel - TCP
-        //ServerSocketChannel - listen all TCP-connections
-    }
-
-    @Test
-    void testBuffer0() throws IOException {
-        byte arr[] = {10, 20, 30, 40, 50};
+    void testBuffer0() {
+        byte[] arr = {10, 20, 30, 40, 50};
         ByteBuffer buffer = ByteBuffer.wrap(arr);
 
         System.out.println("capacity: " + buffer.capacity());
@@ -137,8 +28,8 @@ class MainTest {
     }
 
     @Test
-    void testBuffer1() throws IOException {
-        byte arr[] = {10, 20, 30, 40, 50};
+    void testBuffer1() {
+        byte[] arr = {10, 20, 30, 40, 50};
         ByteBuffer buffer = ByteBuffer.wrap(arr);
 
         System.out.println();
@@ -157,8 +48,8 @@ class MainTest {
     }
 
     @Test
-    void testBuffer2() throws IOException {
-        byte arr[] = {10, 20, 30, 40, 50};
+    void testBuffer2() {
+        byte[] arr = {10, 20, 30, 40, 50};
         ByteBuffer buffer = ByteBuffer.wrap(arr);
 
         System.out.println("get: " + buffer.get());
@@ -173,8 +64,8 @@ class MainTest {
     }
 
     @Test
-    void testBuffer3() throws IOException {
-        byte arr[] = {10, 20, 30, 40, 50};
+    void testBuffer3() {
+        byte[] arr = {10, 20, 30, 40, 50};
         ByteBuffer buffer = ByteBuffer.wrap(arr);
 
         System.out.println("get: " + buffer.get());
@@ -192,8 +83,8 @@ class MainTest {
     }
 
     @Test
-    void testBuffer4() throws IOException {
-        byte arr[] = {10, 20, 30, 40, 50};
+    void testBuffer4() {
+        byte[] arr = {10, 20, 30, 40, 50};
         ByteBuffer buffer = ByteBuffer.wrap(arr);
 
         System.out.println(buffer.get());
@@ -207,8 +98,8 @@ class MainTest {
     }
 
     @Test
-    void testBuffer5() throws IOException {
-        byte arr[] = {10, 20, 30, 40, 50};
+    void testBuffer5() {
+        byte[] arr = {10, 20, 30, 40, 50};
         ByteBuffer buffer = ByteBuffer.wrap(arr);
 
         System.out.println(buffer.get());
@@ -247,7 +138,7 @@ class MainTest {
 
     @Test
     void testWrite() throws IOException {
-        Path destPath = Path.of("channel_example_3.txt");
+        Path destPath = Path.of("dir/channel_example_3.txt");
 
         RandomAccessFile dest = new RandomAccessFile(destPath.toFile(), "rw");
         FileChannel destChannel = dest.getChannel();
@@ -266,19 +157,4 @@ class MainTest {
         destChannel.close();
     }
 
-    @Test
-    void testTruncate() throws IOException {
-        Path path = Path.of("channel_example_4.txt");
-
-        RandomAccessFile file = new RandomAccessFile(path.toFile(), "rw");
-        FileChannel channel = file.getChannel();
-
-        System.out.println(channel.position());
-        System.out.println();
-        System.out.println(channel.size());
-
-        channel.truncate(channel.size() - 20);
-        System.out.println(channel.size());
-        channel.close();
-    }
 }
